+++ pkg/util/util.go
@@ -52,13 +52,6 @@ func GetOperatorNamespaceFromEnv() string {
 	return os.Getenv(OperatorNamespaceEnv)
 }
 
-func CheckPasstImagesEnvExists() error {
-	if _, passtImageVarExists := os.LookupEnv(PasstImageEnvV); !passtImageVarExists {
-		return fmt.Errorf("%s env var not found", PasstImageEnvV)
-	}
-	return nil
-}
-
 func IsRunModeLocal() bool {
 	return os.Getenv(ForceRunModeEnv) == string(LocalRunMode)
 }
